+++ src/playground/App.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { Router, Switch, Route } from 'react-router';
 import { createBrowserHistory } from 'history';
 import ActorEditor from './actor-editor';
+import ActorTemplatePreviewer from './preview';
 
 const history = createBrowserHistory({
   basename: '/playground',
@@ -12,6 +13,7 @@ const App: React.FC = React.memo((props) => {
     <Router history={history}>
       <Switch>
         <Route exact path="/" component={ActorEditor} />
+        <Route exact path="/preview" component={ActorTemplatePreviewer} />
       </Switch>
     </Router>
   );